+++ src/Form/LegacyConsumer/templates/file.html.php
@@ -2,8 +2,9 @@
 <?php /** @var string $fieldIdAttribute */ ?>
 <input
 	type="file"
-	name="<?php echo $field->getName(); ?>"
+	name="<?php echo $field->getName(); ?><?php echo $field->getAllowMultiple() ? '[]' : ''; ?>"
 	id="<?php echo $fieldIdAttribute; ?>"
+	<?php echo $field->getAllowMultiple() ? 'multiple' : ''; ?>
 	<?php echo $field->isRequired() ? 'required' : ''; ?>
 	<?php echo $field->isReadOnly() ? 'readonly' : ''; ?>
 >